Move attack management to a new ActiveScanner class
