+++ wapitiCore/controller/exceptions.py
@@ -1,0 +1,27 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+# This file is part of the Wapiti project (https://wapiti-scanner.github.io)
+# Copyright (C) 2025 Nicolas SURRIBAS
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+
+class InvalidOptionValue(Exception):
+    def __init__(self, opt_name, opt_value):
+        super().__init__()
+        self.opt_name = opt_name
+        self.opt_value = opt_value
+
+    def __str__(self):
+        return f"Invalid argument for option {self.opt_name} : {self.opt_value}"
